+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Dependency to monocle optics library.
+
+### Changed
+
+- Update bifrost node for integration tests to `bifrost-node:2.0.0-alpha7`.
+- The query-utxo command now supports a new parameter to filter UTXOs by token
+types. The valid token types are "lvl", "topl", "asset" and "all". Documentation
+was also updated.
+
+
 ### Removed
 
 - Removed unnecessary validations. Sometimes we were valiting that a parameter